Breaking audience methods out to common package
